+++ chain/chain/src/types.rs
@@ -254,6 +254,8 @@ pub trait RuntimeAdapter: Send + Sync {
         gas_price: Balance,
         state_root: Option<StateRoot>,
         transaction: &SignedTransaction,
+        verify_signature: bool,
+        current_protocol_version: ProtocolVersion,
     ) -> Result<Option<InvalidTxError>, Error>;
 
     /// Returns an ordered list of valid transactions from the pool up the given limits.
@@ -271,6 +273,7 @@ pub trait RuntimeAdapter: Send + Sync {
         state_root: StateRoot,
         pool_iterator: &mut dyn PoolIterator,
         chain_validate: &mut dyn FnMut(&SignedTransaction) -> bool,
+        current_protocol_version: ProtocolVersion,
     ) -> Result<Vec<SignedTransaction>, Error>;
 
     /// Verify validator signature for the given epoch.
@@ -566,14 +569,14 @@ pub trait RuntimeAdapter: Send + Sync {
                 .cloned()
                 .collect();
             receipts_hashes
-                .push(hash(&ReceiptList(shard_id, shard_receipts).try_to_vec().unwrap()));
+                .push(hash(&ReceiptList(shard_id, &shard_receipts).try_to_vec().unwrap()));
         }
         receipts_hashes
     }
 }
 
-#[derive(BorshSerialize, BorshDeserialize, Serialize, Debug, Clone, Default)]
-pub struct ReceiptList(pub ShardId, pub Vec<Receipt>);
+#[derive(BorshSerialize, Serialize, Debug, Clone)]
+pub struct ReceiptList<'a>(pub ShardId, pub &'a Vec<Receipt>);
 
 /// The last known / checked height and time when we have processed it.
 /// Required to keep track of skipped blocks and not fallback to produce blocks at lower height.
@@ -581,17 +584,6 @@ pub struct ReceiptList(pub ShardId, pub Vec<Receipt>);
 pub struct LatestKnown {
     pub height: BlockHeight,
     pub seen: u64,
-}
-
-/// When running block sync response to know if the node needs to sync state,
-/// or the hashes from the blocks that are needed.
-pub enum BlockSyncResponse {
-    /// State is needed before we start fetching recent blocks.
-    StateNeeded,
-    /// We are up to date with state, list of block hashes that need to be fetched.
-    BlocksNeeded(Vec<CryptoHash>),
-    /// We are up to date, nothing is required.
-    None,
 }
 
 #[cfg(test)]