+++ runtime/runtime/src/verifier.rs
@@ -13,6 +13,7 @@ use near_primitives::transaction::{
     SignedTransaction, StakeAction,
 };
 use near_primitives::utils::is_valid_account_id;
+use near_primitives::version::ProtocolVersion;
 use near_store::{get_access_key, get_account, set_access_key, set_account, TrieUpdate};
 use near_vm_logic::types::Balance;
 use near_vm_logic::VMLimitConfig;
@@ -23,6 +24,8 @@ pub fn validate_transaction(
     config: &RuntimeConfig,
     gas_price: Balance,
     signed_transaction: &SignedTransaction,
+    verify_signature: bool,
+    current_protocol_version: ProtocolVersion,
 ) -> Result<TransactionCost, RuntimeError> {
     let transaction = &signed_transaction.transaction;
     let signer_id = &transaction.signer_id;
@@ -36,9 +39,10 @@ pub fn validate_transaction(
         .into());
     }
 
-    if !signed_transaction
-        .signature
-        .verify(signed_transaction.get_hash().as_ref(), &transaction.public_key)
+    if verify_signature
+        && !signed_transaction
+            .signature
+            .verify(signed_transaction.get_hash().as_ref(), &transaction.public_key)
     {
         return Err(InvalidTxError::InvalidSignature.into());
     }
@@ -48,8 +52,14 @@ pub fn validate_transaction(
 
     let sender_is_receiver = &transaction.receiver_id == signer_id;
 
-    tx_cost(&config.transaction_costs, &transaction, gas_price, sender_is_receiver)
-        .map_err(|_| InvalidTxError::CostOverflow.into())
+    tx_cost(
+        &config.transaction_costs,
+        &transaction,
+        gas_price,
+        sender_is_receiver,
+        current_protocol_version,
+    )
+    .map_err(|_| InvalidTxError::CostOverflow.into())
 }
 
 /// Verifies the signed transaction on top of given state, charges transaction fees
@@ -59,9 +69,17 @@ pub fn verify_and_charge_transaction(
     state_update: &mut TrieUpdate,
     gas_price: Balance,
     signed_transaction: &SignedTransaction,
+    verify_signature: bool,
+    current_protocol_version: ProtocolVersion,
 ) -> Result<VerificationResult, RuntimeError> {
     let TransactionCost { gas_burnt, gas_remaining, receipt_gas_price, total_cost, burnt_amount } =
-        validate_transaction(config, gas_price, signed_transaction)?;
+        validate_transaction(
+            config,
+            gas_price,
+            signed_transaction,
+            verify_signature,
+            current_protocol_version,
+        )?;
     let transaction = &signed_transaction.transaction;
     let signer_id = &transaction.signer_id;
 
@@ -414,6 +432,7 @@ mod tests {
         CreateAccountAction, DeleteKeyAction, StakeAction, TransferAction,
     };
     use near_primitives::types::{AccountId, Balance, MerkleHash, StateChangeCause};
+    use near_primitives::version::PROTOCOL_VERSION;
     use near_store::test_utils::create_tries;
     use std::convert::TryInto;
     use std::sync::Arc;
@@ -473,13 +492,20 @@ mod tests {
         expected_err: RuntimeError,
     ) {
         assert_eq!(
-            validate_transaction(&config, gas_price, &signed_transaction)
+            validate_transaction(&config, gas_price, &signed_transaction, true, PROTOCOL_VERSION)
                 .expect_err("expected an error"),
             expected_err,
         );
         assert_eq!(
-            verify_and_charge_transaction(&config, state_update, gas_price, &signed_transaction)
-                .expect_err("expected an error"),
+            verify_and_charge_transaction(
+                &config,
+                state_update,
+                gas_price,
+                &signed_transaction,
+                true,
+                PROTOCOL_VERSION,
+            )
+            .expect_err("expected an error"),
             expected_err,
         );
     }
@@ -501,10 +527,17 @@ mod tests {
             deposit,
             CryptoHash::default(),
         );
-        validate_transaction(&config, gas_price, &transaction).expect("valid transaction");
-        let verification_result =
-            verify_and_charge_transaction(&config, &mut state_update, gas_price, &transaction)
-                .expect("valid transaction");
+        validate_transaction(&config, gas_price, &transaction, true, PROTOCOL_VERSION)
+            .expect("valid transaction");
+        let verification_result = verify_and_charge_transaction(
+            &config,
+            &mut state_update,
+            gas_price,
+            &transaction,
+            true,
+            PROTOCOL_VERSION,
+        )
+        .expect("valid transaction");
         // Should not be free. Burning for sending
         assert!(verification_result.gas_burnt > 0);
         // All burned gas goes to the validators at current gas price
@@ -622,6 +655,8 @@ mod tests {
                     100,
                     CryptoHash::default(),
                 ),
+                false,
+                PROTOCOL_VERSION,
             )
             .expect_err("expected an error"),
             RuntimeError::InvalidTxError(InvalidTxError::InvalidAccessKeyError(
@@ -686,6 +721,8 @@ mod tests {
                     100,
                     CryptoHash::default(),
                 ),
+                true,
+                PROTOCOL_VERSION,
             )
             .expect_err("expected an error"),
             RuntimeError::InvalidTxError(InvalidTxError::SignerDoesNotExist {
@@ -716,6 +753,8 @@ mod tests {
                     100,
                     CryptoHash::default(),
                 ),
+                true,
+                PROTOCOL_VERSION,
             )
             .expect_err("expected an error"),
             RuntimeError::InvalidTxError(InvalidTxError::InvalidNonce { tx_nonce: 1, ak_nonce: 2 }),
@@ -762,6 +801,8 @@ mod tests {
                 TESTING_INIT_BALANCE,
                 CryptoHash::default(),
             ),
+            true,
+            PROTOCOL_VERSION,
         )
         .expect_err("expected an error");
         if let RuntimeError::InvalidTxError(InvalidTxError::NotEnoughBalance {
@@ -811,6 +852,8 @@ mod tests {
                 })],
                 CryptoHash::default(),
             ),
+            true,
+            PROTOCOL_VERSION,
         )
         .expect_err("expected an error");
         if let RuntimeError::InvalidTxError(InvalidTxError::InvalidAccessKeyError(
@@ -850,6 +893,8 @@ mod tests {
                     transfer_amount,
                     CryptoHash::default(),
                 ),
+                true,
+                PROTOCOL_VERSION,
             )
             .expect_err("expected an error"),
             RuntimeError::InvalidTxError(InvalidTxError::LackBalanceForState {
@@ -898,6 +943,8 @@ mod tests {
                     ],
                     CryptoHash::default(),
                 ),
+                true,
+                PROTOCOL_VERSION,
             )
             .expect_err("expected an error"),
             RuntimeError::InvalidTxError(InvalidTxError::InvalidAccessKeyError(
@@ -918,6 +965,8 @@ mod tests {
                     vec![],
                     CryptoHash::default(),
                 ),
+                true,
+                PROTOCOL_VERSION,
             )
             .expect_err("expected an error"),
             RuntimeError::InvalidTxError(InvalidTxError::InvalidAccessKeyError(
@@ -938,6 +987,8 @@ mod tests {
                     vec![Action::CreateAccount(CreateAccountAction {})],
                     CryptoHash::default(),
                 ),
+                true,
+                PROTOCOL_VERSION,
             )
             .expect_err("expected an error"),
             RuntimeError::InvalidTxError(InvalidTxError::InvalidAccessKeyError(
@@ -980,6 +1031,8 @@ mod tests {
                     }),],
                     CryptoHash::default(),
                 ),
+                true,
+                PROTOCOL_VERSION,
             )
             .expect_err("expected an error"),
             RuntimeError::InvalidTxError(InvalidTxError::InvalidAccessKeyError(
@@ -1025,6 +1078,8 @@ mod tests {
                     }),],
                     CryptoHash::default(),
                 ),
+                true,
+                PROTOCOL_VERSION,
             )
             .expect_err("expected an error"),
             RuntimeError::InvalidTxError(InvalidTxError::InvalidAccessKeyError(
@@ -1067,6 +1122,8 @@ mod tests {
                     }),],
                     CryptoHash::default(),
                 ),
+                true,
+                PROTOCOL_VERSION,
             )
             .expect_err("expected an error"),
             RuntimeError::InvalidTxError(InvalidTxError::InvalidAccessKeyError(